split out the test for amd, and fix comments / headers

+++ salsa20/salsa/salsa20_noasm.go
@@ -13,6 +13,6 @@ func XORKeyStream(out, in []byte, counter *[16]byte, key *[32]byte) {
 	genericXORKeyStream(out, in, counter, key)
 }
 
-func XORKeyStreamWithRounds(out, in []byte, counter *[16]byte, key *[32]byte, rounds uint64) {
+func XORKeyStreamWithRounds(out, in []byte, counter *[16]byte, key *[32]byte, rounds uint) {
 	generic20nXORKeyStream(out, in, counter, key, rounds)
 }
